Add vertex ambiant occlusion

+++ app/include/MeshBuilder.h
@@ -28,12 +28,12 @@ class MeshBuilder {
     void reset();
     void addCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<float, 4>& ao, Face face);
     
-    void addFrontCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<float, 4>& ao);
-    void addBackCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<float, 4>& ao);
-    void addRightCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<float, 4>& ao);
-    void addLeftCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<float, 4>& ao);
-    void addTopCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<float, 4>& ao);
-    void addBottomCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<float, 4>& ao);
+    void addFrontCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<int, 4>& ao);
+    void addBackCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<int, 4>& ao);
+    void addRightCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<int, 4>& ao);
+    void addLeftCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<int, 4>& ao);
+    void addTopCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<int, 4>& ao);
+    void addBottomCubeFace(const glm::vec3& pos, const opticrafter::UVRegion& region, std::array<int, 4>& ao);
 
     std::unique_ptr<opticrafter::Mesh> build();
